# openapi.yaml
openapi: 3.1.0
info:
  title: Campaign and Ad Management API
  version: 1.0.0
  description: API for managing users, sponsors, influencers, ad requests, and campaigns.

# The servers section defines the API base URL
servers:
  - url: /api
paths:
...
  /campaigns:
    get:
      summary: Get all campaigns
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
        '500':
          description: Error fetching campaigns
    post:
      summary: Create a new campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Campaign'
      responses:
        '201':
          description: Campaign created successfully
        '400':
          description: Invalid input
        '500':
          description: Failed to create campaign

  # Endpoint for a specific campaign by campaign_id
...
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        campaign_id:
          type: integer
        name:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        budget:
          type: number
          format: float